+++ CHANGELOG.md
@@ -2,11 +2,19 @@ Change Log
 ==========
 
 ## [Unreleased]
+* [Issue #5](https://github.com/dSpaceLabs/Queue/issues/5): Support for AWS SQS FIFO
+* Added RabbitQueue
+* Bug fix with Redis Queue Name not being set
+
+## [0.2.0]
+* Ability to now publish a message without wrapping it in a Message object
+* Added RedisQueue
 
 
 ## [0.1.0]
 * Initial Release
 * Includes Sqs and File queues
 
-[Unreleased]: https://github.com/dSpaceLabs/Queue/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/dSpaceLabs/Queue/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/dSpaceLabs/Queue/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dSpaceLabs/Queue/compare/e30e9cdec4368d13425509761b8dcc9fbdaea517...v0.1.0